auth update

+++ quotes-backend/Controllers/QuotesController.cs
@@ -3,6 +3,7 @@ using System.Security.Claims;
 using DTO;
 using Microsoft.AspNetCore.Mvc;
 using Logic;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Swashbuckle.AspNetCore.Annotations;
 
@@ -18,15 +19,15 @@ public class QuotesController : ControllerBase
         _quoteCollection = new QuoteCollection();
     }
     [SwaggerOperation(
-        Summary = "Gets random Quote",
-        Description = "Requires AUTH for retrieving context"
+        Summary = "Gets random Quote (requires AUTH)",
+        Description = "Requires special permissions to retrieve the context"
     )]
     [HttpGet]
     [Route("Random")]
     public IActionResult GetRandom(bool withContext, bool asObject)
     {
         // Check if either parameter requires permission
-        if ((withContext) && (!User.Identity.IsAuthenticated || !User.HasClaim(c => c.Type == "Rights" && c.Value == "True")))
+        if ((withContext) && (!Rights.hasRights(User)))
         {
             return Unauthorized("Access denied: please log in with appropriate permissions.");
         }
@@ -44,8 +45,8 @@ public class QuotesController : ControllerBase
     }
 
     [SwaggerOperation(
-        Summary = "Gets all quotes from database",
-        Description = "When the text field is filled in, the API will try to find the requested quote. Only works with AUTH"
+        Summary = "Gets all quotes from database (requires AUTH)",
+        Description = "When the text field is filled in, the API will try to find the requested quote. Only works with the correct rights"
     )]
     [HttpGet]
     [Route("/Quotes")]
@@ -57,7 +58,7 @@ public class QuotesController : ControllerBase
         }
         List<Quote> allquotes = new List<Quote>();
         if(text is not null){
-            if (User.Identity.IsAuthenticated && User.HasClaim(c => c.Type == "Rights" && c.Value == "True"))
+            if (Rights.hasRights(User))
             {
                 allquotes.Add(_quoteCollection.FindQuoteBasedOnText(text));
                 if (allquotes.Count != 0)
@@ -69,7 +70,7 @@ public class QuotesController : ControllerBase
             return Unauthorized("Please kys or login to retrieve the context");
         }
         if(context is not null){
-            if (User.Identity.IsAuthenticated && User.HasClaim(c => c.Type == "Rights" && c.Value == "True"))
+            if (Rights.hasRights(User))
             {
                 allquotes.Add(_quoteCollection.FindQuoteBasedOnContext(context));
                 if (allquotes.Count != 0 && allquotes[0].text is not null)
@@ -80,7 +81,7 @@ public class QuotesController : ControllerBase
             }
             return Unauthorized("Please kys or login to retrieve the context");
         }
-        if (User.Identity.IsAuthenticated && User.HasClaim(c => c.Type == "Rights" && c.Value == "True"))
+        if (Rights.hasRights(User))
         {
            allquotes = _quoteCollection.GetAllQuotes(true);
         }
@@ -95,9 +96,9 @@ public class QuotesController : ControllerBase
         return NoContent();
     }
     [SwaggerOperation(
-        Summary = "Adds a new Quote",
-        Description = "Requires AUTH"
-    )]
+        Summary = "Adds a new Quote (requires AUTH)",
+        Description = "Requires special permissions to create a new quote"
+        )]
     [HttpPost(Name = "NewQuote"), Authorize]
     public IActionResult NewQuote([FromBody] QuoteDTOPost quote)
     {
@@ -110,19 +111,17 @@ public class QuotesController : ControllerBase
         return BadRequest();
     }
     [SwaggerOperation(
-        Summary = "Edits a quote",
-        Description = "Requires AUTH"
+        Summary = "Edits a quote (requires AUTH)",
+        Description = "Requires special permissions to update"
     )]
     [HttpPut]
     [Route("{id}"), Authorize]
     public IActionResult Update(string id, [FromBody] Quote quote)
     {
         bool updated = false;
-        if (User.Identity.IsAuthenticated && User.HasClaim(c => c.Type == "Rights" && c.Value == "True"))
+        if (Rights.hasRights(User))
         { 
-            updated = _quoteCollection.UpdateQuote(id, quote, true);
-        }else{
-            updated = _quoteCollection.UpdateQuote(id, quote, false);
+            updated = _quoteCollection.UpdateQuote(id, quote);
         }
         
         if (updated)
@@ -130,23 +129,24 @@ public class QuotesController : ControllerBase
             return Ok();
         }
 
-        return BadRequest();
+        return Forbid();
 
     }
     [SwaggerOperation(
-        Summary = "Deletes a Quote",
-        Description = "Requires AUTH"
+        Summary = "Deletes a Quote (requires AUTH)",
+    Description = "Requires special permissions to delete"
     )]
     [HttpDelete]
     [Route("{id}"),Authorize]
     public IActionResult Delete(string id)
     {
-        bool deleted = _quoteCollection.DeleteQuote(id);
-        if (deleted)
-        {
-            return Ok();
+        if (Rights.hasRights(User)){
+            bool deleted = _quoteCollection.DeleteQuote(id);
+            if (deleted)
+            {
+                return Ok();
+            }
         }
-
-        return BadRequest();
+        return Forbid();
     }
 }
